Update synopsys references to blackduck
