+++ README.md
@@ -1,8 +1,8 @@
 # Overview
-Synopsys Detect for Jenkins simplifies running Synopsys Detect in your Jenkins builds.
+Detect for Jenkins simplifies running Detect in your Jenkins builds.
 
 # Build
-[![Build Status](https://travis-ci.org/jenkinsci/synopsys-detect-plugin.svg?branch=master)](https://travis-ci.org/jenkinsci/synopsys-detect-plugin)
+[![Build Status](https://travis-ci.org/jenkinsci/blackduck-detect-plugin.svg?branch=master)](https://travis-ci.org/jenkinsci/blackduck-detect-plugin)
 [![Quality Gate Status](https://sonarcloud.io/api/project_badges/measure?project=com.blackducksoftware.integration%3Ablackduck-detect&metric=alert_status)](https://sonarcloud.io/dashboard?id=com.blackducksoftware.integration%3Ablackduck-detect)
 [![Coverage Status](https://coveralls.io/repos/github/jenkinsci/synopsys-detect-plugin/badge.svg?branch=master)](https://coveralls.io/github/jenkinsci/synopsys-detect-plugin?branch=master)
 [![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](https://opensource.org/licenses/Apache-2.0) 
@@ -11,4 +11,4 @@ Synopsys Detect for Jenkins simplifies running Synopsys Detect in your Jenkins b
 You can download the latest artifact from releases: https://updates.jenkins-ci.org/download/plugins/blackduck-detect/
 
 # Documentation
-Documentation for Synopsys Detect for Jenkins can be found [here](https://sig-product-docs.synopsys.com/bundle/integrations-detect/page/integrations/jenkinsplugin/jenkins.html)
+Documentation for Detect for Jenkins can be found [here](https://sig-product-docs.synopsys.com/bundle/integrations-detect/page/integrations/jenkinsplugin/jenkins.html)
