+++ build.gradle
@@ -47,7 +47,7 @@ project.ext.excludesFromTestCoverage = ['**/DetectDownloadStrategy.java', '**/De
 
 group = 'com.blackducksoftware.integration'
 version = '9.0.0-SNAPSHOT'
-description = 'Synopsys Detect Plugin for Jenkins'
+description = 'Detect Plugin for Jenkins'
 
 apply plugin: 'com.synopsys.integration.solution'
 
@@ -59,7 +59,7 @@ artifactory {
 
 jenkinsPlugin {
     jenkinsVersion = '2.401.3' // as of version 0.40.0 of jpi, it's jenkinsVersion...
-    displayName = 'Synopsys Detect plugin'
+    displayName = 'Detect plugin'
     compatibleSinceVersion = '9.0.0'
     url = 'https://wiki.jenkins.io/display/JENKINS/Synopsys+Detect+Plugin'
     gitHubUrl = 'https://github.com/jenkinsci/synopsys-detect-plugin'
@@ -69,10 +69,10 @@ jenkinsPlugin {
 
 dependencies {
     implementation 'org.springframework.security:spring-security-web:5.8.5'
-    annotationProcessor 'com.synopsys.integration:jenkins-annotation-processor:0.0.7'
+    annotationProcessor 'com.blackduck.integration:jenkins-annotation-processor:0.0.7'
 
-    implementation 'com.synopsys.integration:blackduck-common:66.1.2'
-    implementation ('com.synopsys.integration:jenkins-common:0.5.5') {
+    implementation 'com.blackduck.integration:blackduck-common:66.1.2'
+    implementation ('com.blackduck.integration:jenkins-common:0.5.5') {
         exclude group: "org.jenkins-ci.main", module: "jenkins-core"
         exclude group: "org.jenkins-ci.plugins", module: "credentials"
         exclude group: "org.jenkins-ci.plugins", module: "plain-credentials"
