+++ src/main/java/com/blackduck/integration/detect/DetectJenkinsEnvironmentVariable.java
@@ -1,0 +1,24 @@
+/*
+ * blackduck-detect
+ *
+ * Copyright (c) 2023 Synopsys, Inc.
+ *
+ * Use subject to the terms and conditions of the Synopsys End User Software License and Maintenance Agreement. All rights reserved worldwide.
+ */
+package com.blackduck.integration.detect;
+
+public enum DetectJenkinsEnvironmentVariable {
+    USER_PROVIDED_JAR_PATH("DETECT_JAR"),
+    SHOULD_ESCAPE("DETECT_PLUGIN_ESCAPING");
+
+    private final String environmentVariable;
+
+    DetectJenkinsEnvironmentVariable(String environmentVariable) {
+        this.environmentVariable = environmentVariable;
+    }
+
+    public String stringValue() {
+        return environmentVariable;
+    }
+
+}
