+++ src/main/resources/com/blackduck/integraton/jenkins/detect/extensions/AirGapDownloadStrategy/config.jelly
