+++ src/main/resources/com/blackduck/integraton/jenkins/detect/extensions/DetectGlobalConfig/config.jelly
