+++ src/main/resources/com/blackduck/integraton/jenkins/detect/extensions/DetectPipelineStep/config.jelly
