+++ src/main/resources/com/blackduck/integraton/jenkins/detect/extensions/InheritFromGlobalDownloadStrategy/config.jelly
