+++ src/main/resources/com/blackduck/integraton/jenkins/detect/extensions/ScriptOrJarDownloadStrategy/config.jelly
